// src/types/StatusDocument.ts
export interface StatusDocument {
  DocumentWeight: number | string;
  ServiceType: string;
  UndeliveryReasons: string;
  RecipientFullName: string;
  MarketplacePartnerToken: string;
  CounterpartySenderDescription: string;
  InternationalDeliveryType: 'Import' | '';
  CargoType: string;
  PayerType: 'Sender' | 'Recipient';
  SeatsAmount: string;
  PhoneRecipient: string;
  ScheduledDeliveryDate: string;
  DocumentCost: string;
  CardMaskedNumber: string;
  OwnerDocumentType: string;
  ExpressWaybillPaymentStatus: 'NeedPayment' | 'Payed';
  ExpressWaybillAmountToPay?: number;
  AfterpaymentOnGoodsCost: number;
  SumBeforeCheckWeight: number;
  CheckWeight: number;
  PaymentMethod: string;
  AdjustedDate: string;
  Number: string;
  PhoneSender: string;
  CalculatedWeight: string;
  WarehouseRecipient: string;
  WarehouseSender: string;
  DateCreated: string;
  DateScan: string;
  DateReturnCargo?: string;
  DateMoving: string;
  RecipientAddress: string;
  RecipientDateTime: string;
  RecipientFullNameEW: string;
  RefCityRecipient: string;
  RefCitySender: string;
  RefSettlementRecipient: string;
  RefSettlementSender: string;
  SenderAddress: string;
  SenderFullNameEW: string;
  ClientBarcode: string;
  CitySender: string;
  CityRecipient: string;
  CargoDescriptionString: string;
  AnnouncedPrice: number;
  AdditionalInformationEW: string;
  ActualDeliveryDate: string;
  StatusCode: string;
  PostomatV3CellReservationNumber: boolean;
  AmountToPay: string;
  AmountPaid: string;
  VolumeWeight: string;
  CheckWeightMethod: string;
  OwnerDocumentNumber: string;
  LastCreatedOnTheBasisNumber: string;
  LastCreatedOnTheBasisDateTime: string;
  LastTransactionDateTimeGM: string;
  PaymentStatus: string;
  PaymentStatusDate: string;
  LastAmountTransferGM: string;
  LastAmountReceivedCommissionGM: number;
  LastCreatedOnTheBasisPayerType: string;
  DeliveryTimeframe: string;
  LastTransactionStatusGM: string;
  Status: string;
  Redelivery: number;
  RedeliveryNum: string;
  RedeliverySum: number;
  RedeliveryPayer: string;
  UndeliveryReasonsDate: string;
  UndeliveryReasonsSubtypeDescription: string;
  RecipientWarehouseTypeRef: string;
  WarehouseRecipientInternetAddressRef: string;
  WarehouseRecipientNumber: number;
  WarehouseRecipientRef: string;
  CategoryOfWarehouse: string;
  WarehouseRecipientAddress: string;
  WarehouseSenderInternetAddressRef: string;
  WarehouseSenderAddress: string;
  CounterpartyType: string;
  CounterpartySenderType: string;
  AviaDelivery: number;
  BarcodeRedBox: string;
  CargoReturnRefusal: boolean;
  DaysStorageCargo: string;
  Packaging: unknown[]; // todo
  PartialReturnGoods: unknown[]; // todo
  SecurePayment: boolean;
  PossibilityCreateRedirecting: boolean;
  PossibilityCreateReturn: boolean;
  PossibilityCreateRefusal: boolean;
  PossibilityChangeEW: boolean;
  PossibilityChangeDeliveryIntervals: boolean;
  PossibilityTrusteeRecipient: boolean;
  TrusteeRecipientPhone: string;
  CounterpartyRecipientDescription: string;
  CreatedOnTheBasis: string;
  FactualWeight: string;
  FreeShipping: string;
  InternetDocumentDescription: string;
  LastCreatedOnTheBasisDocumentType: string;
  LoyaltyCardRecipient: string;
  LoyaltyCardSender: string;
  RedeliveryPaymentCardDescription: string;
  RedeliveryServiceCost: string;
  RedeliveryPaymentCardRef: string;
  PossibilityLightReturn: string;
  RefEW: string;
}